// project/back/app/src/utils/requestcode.enum.ts
export enum FriendCode {
  FRIEND_REQUEST_SENT = 0,
  UNEXISTING_USER = 1,
  NEW_FRIEND = 2,
  ALREADY_FRIEND = 3,
  FRIEND_REQUEST = 4,
  UNFRIEND_SUCCESS = 5,
  UNEXISTING_FRIEND = 6,
  NEW_UNFRIEND = 7,
}

export enum messageCode {
  SUCCESS = 0,
  UNEXISTING_CHANNEL = 1,
  UNACCESSIBLE_CHANNEL = 2,
  INVALID_FORMAT = 3,
  UNAUTHORIZED = 401,
}

export enum ChannelCode {
  INVITE_SUCCESS = 0,
  UNEXISTING_USER = 1,
  UNEXISTING_CHANNEL = 2,
}